endpoints/requirements.txt
acres==0.2.0
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
dnspython==2.7.0
etelemetry==0.3.1
fastapi==0.115.8
filelock==3.17.0
fitz==0.0.1.dev2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.6.1
jiter==0.8.2
looseversion==1.3.0
lxml==5.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pymongo==4.11
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyxnat==1.6.3
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
scipy==1.15.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tabulate==0.9.0
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
